@@ -1,11 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+
+    compileSdkVersion 29
+    ndkVersion '23.0.7123448 rc1'
+
     defaultConfig {
         applicationId "com.actor.testapplication"
         minSdkVersion 21
-        targetSdkVersion 28
+        targetSdkVersion 29
         versionCode 19102101
         versionName "1.1"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -95,19 +98,16 @@ dependencies {
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
 
-//    implementation 'com.android.support:design:28.0.0'
-//    implementation 'com.android.support:support-v4:28.0.0'
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 
     //https://github.com/JakeWharton/butterknife
-    implementation 'com.jakewharton:butterknife:8.8.1'
-    annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.1'
+    implementation 'com.jakewharton:butterknife:10.2.3'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:10.2.3'
 
     //https://github.com/actor20170211030627/MyAndroidFrameWork
-    implementation 'com.github.actor20170211030627.MyAndroidFrameWork:myandroidframework:1.4.1'
+    implementation 'com.github.actor20170211030627.MyAndroidFrameWork:myandroidframework:2.0.1'
 
-    //Okhttp3的Cookie
+    //https://github.com/franmontiel/PersistentCookieJar Okhttp3的Cookie
     implementation 'com.github.franmontiel:PersistentCookieJar:v1.0.1'
 
     //https://github.com/ReactiveX/RxJava
