@@ -1,25 +1,19 @@
 package com.actor.testapplication;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.actor.myandroidframework.application.ActorApplication;
 import com.actor.myandroidframework.utils.LogUtils;
 import com.actor.testapplication.utils.AssetsUtils;
-import com.actor.testapplication.utils.GsonField;
-import com.blankj.utilcode.util.GsonUtils;
 import com.franmontiel.persistentcookiejar.ClearableCookieJar;
 import com.franmontiel.persistentcookiejar.PersistentCookieJar;
 import com.franmontiel.persistentcookiejar.cache.SetCookieCache;
 import com.franmontiel.persistentcookiejar.persistence.SharedPrefsCookiePersistor;
-import com.google.gson.annotations.SerializedName;
-import com.zhy.http.okhttp.OkHttpUtils;
 import com.zhy.http.okhttp.https.HttpsUtils;
-import com.zhy.http.okhttp.log.LoggerInterceptor;
 
 import java.io.InputStream;
 import java.security.cert.Certificate;
-import java.util.Date;
 
 import javax.net.ssl.HostnameVerifier;
 import javax.net.ssl.SSLPeerUnverifiedException;
@@ -37,8 +31,6 @@ import okio.Buffer;
  * @version 1.0
  */
 public class MyApplication extends ActorApplication {
-
-    public static MyApplication instance;
 
     //使用字符串替代服务器.cer证书
     private static final String CER_SERVER = "-----BEGIN CERTIFICATE-----\n" +
@@ -63,20 +55,6 @@ public class MyApplication extends ActorApplication {
     @Nullable
     @Override
     protected OkHttpClient.Builder configOkHttpClientBuilder(OkHttpClient.Builder builder) {
-        return null;
-    }
-
-    @NonNull
-    @Override
-    protected String getBaseUrl() {
-        return "https://api.github.com";
-    }
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        instance = this;
-
         //持久化Cookie & Session框架:PersistentCookieJar
         ClearableCookieJar cookieJar = new PersistentCookieJar(new SetCookieCache(), new SharedPrefsCookiePersistor(this));
         /**
@@ -91,9 +69,10 @@ public class MyApplication extends ActorApplication {
         InputStream bksFile = AssetsUtils.openFile(this, "zhy_client.bks");//使用zhy_client.jks要报错?
         InputStream[] certificates = {certificate};
         HttpsUtils.SSLParams sslParams = HttpsUtils.getSslSocketFactory(certificates, bksFile, "123456");
-        OkHttpClient okHttpClient = new OkHttpClient.Builder()
+
+        return builder
                 //Log
-                .addInterceptor(new LoggerInterceptor(getPackageName() + ",Interceptor:", isDebugMode))//tag, showResponse
+//                .addInterceptor(new LoggerInterceptor(getPackageName() + ",Interceptor:", isAppDebug()))//tag, showResponse
 //                .addNetworkInterceptor(new LoggerInterceptor(getPackageName() + ",NetworkInterceptor:", isDebugMode))
 //                .authenticator(new Authenticator() {//认证x
 //                    @Override
@@ -125,7 +104,7 @@ public class MyApplication extends ActorApplication {
                             //获取证书链中的所有证书
                             Certificate[] peerCertificates = session.getPeerCertificates();
                             for (Certificate c : peerCertificates) {//打印所有证书内容
-                                LogUtils.println("verify: "+c.toString(), true);
+                                LogUtils.error("verify: "+c.toString(), true);
                             }
                         } catch (SSLPeerUnverifiedException e) {
                             e.printStackTrace();
@@ -149,47 +128,17 @@ public class MyApplication extends ActorApplication {
 //                .sslSocketFactory(null)//过时
                 .sslSocketFactory(sslParams.sSLSocketFactory, sslParams.trustManager)//设置具体的证书
 //                .writeTimeout(10_000L, TimeUnit.MILLISECONDS)//写超时,默认10秒
-                .build();
-        OkHttpUtils.initClient(okHttpClient);
-
-
-
-        // TODO: 2020/7/30 未完成
-//        Gson gson = GsonUtils.getGson().newBuilder()
-//                .registerTypeAdapterFactory(DateJsonDeserializer.FACTORY)
-//                .create();
-//        GsonUtils.setGsonDelegate(gson);
-
-        User user = new User(11, new Date(System.currentTimeMillis()));
-        String json = GsonUtils.toJson(user);
-        System.out.println("json: " + json);
-        json = "{\"age\":\"11\",\"date\":\"2020-01-02 03:04:05\"}";//Jul 30, 2020 4:26:16 PM
-
-//        User user1 = GsonUtils.fromJson(json, User.class);
-        System.out.println(1);
+                ;
     }
 
-
+    @NonNull
+    @Override
+    protected String getBaseUrl(boolean isDebugMode) {
+        return "https://api.github.com";
+    }
 
     @Override
     protected void onUncaughtException(Thread thread, Throwable e) {
 //        System.exit(-1);//退出
-    }
-}
-
-class User {
-
-    @SerializedName("age")
-    public int age;
-    //    public Integer age;
-
-    @GsonField
-    public Date date;
-
-    public User() {}
-    public User(int age, Date date) {
-//    public User(Integer age, Date date) {
-        this.age = age;
-        this.date = date;
     }
 }
