@@ -1,7 +1,8 @@
 package com.actor.testapplication.activity;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.Nullable;
 
 import com.actor.testapplication.R;
 
