@@ -1,8 +1,9 @@
 package com.actor.testapplication.activity;
 
 import android.os.Bundle;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.actor.testapplication.R;
 import com.actor.testapplication.bean.ExpandableItemAdapter;
