@@ -1,11 +1,12 @@
 package com.actor.testapplication.widget;
 
 import android.content.Context;
-import android.support.v4.widget.ViewDragHelper;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 import android.view.View;
 import android.widget.FrameLayout;
+
+import androidx.customview.widget.ViewDragHelper;
 
 import com.actor.testapplication.R;
 
